requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
abseil-cpp=20211102.0=hc377ac9_0
aiohttp=3.8.3=py311h80987f9_0
aiosignal=1.2.0=pyhd3eb1b0_0
arrow-cpp=11.0.0=hc7aafb3_1
async-timeout=4.0.2=py311hca03da5_0
attrs=22.1.0=py311hca03da5_0
aws-c-common=0.6.8=h80987f9_1
aws-c-event-stream=0.1.6=h313beb8_6
aws-checksums=0.1.11=h80987f9_2
aws-sdk-cpp=1.8.185=ha71a6ea_1
blas=2.117=openblas
blas-devel=3.9.0=17_osxarm64_openblas
boost-cpp=1.73.0=h1a28f6b_12
bottleneck=1.3.5=py311ha0d4635_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
brotli-python=1.0.9=py311h313beb8_7
bzip2=1.0.8=h620ffc9_4
c-ares=1.19.0=h80987f9_0
ca-certificates=2023.7.22=hf0a4a13_0
catalogue=2.0.7=py311hca03da5_0
certifi=2023.7.22=py311hca03da5_0
cffi=1.15.1=py311h80987f9_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.6=unix_pyh707e725_0
colorama=0.4.6=pyhd8ed1ab_0
commonmark=0.9.1=py_0
confection=0.0.4=py311hb6e6a13_0
cymem=2.0.6=py311h313beb8_0
cython-blis=0.7.9=py311hb9f6ed7_0
dataclasses=0.8=pyhc8e2a94_3
datasets=2.14.4=py_0
dill=0.3.6=py311hca03da5_0
en-core-web-sm=3.5.0=pypi_0
filelock=3.9.0=py311hca03da5_0
freetype=2.12.1=h1192e45_0
frozenlist=1.3.3=py311h80987f9_0
fsspec=2023.4.0=py311hca03da5_0
future=0.18.3=pyhd8ed1ab_0
gflags=2.2.2=hc377ac9_0
giflib=5.2.1=h80987f9_3
glog=0.5.0=hc377ac9_0
gmp=6.2.1=hc377ac9_3
gmpy2=2.1.2=py311h40f64dc_0
grpc-cpp=1.48.2=hc60591f_1
huggingface_hub=0.16.4=py_0
icu=68.1=hc377ac9_0
idna=3.4=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=py311hca03da5_0
jpeg=9e=h80987f9_1
krb5=1.20.1=hf3e1bf2_1
langcodes=3.3.0=pyhd8ed1ab_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libblas=3.9.0=17_osxarm64_openblas
libboost=1.73.0=h49e8a49_12
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcblas=3.9.0=17_osxarm64_openblas
libcurl=8.1.1=h3e2b118_2
libcxx=16.0.6=h4653b0c_0
libdeflate=1.17=h80987f9_0
libedit=3.1.20221030=h80987f9_0
libev=4.33=h1a28f6b_1
libevent=2.1.12=h02f6b3c_1
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=12_3_0_hd922786_1
libgfortran5=12.3.0=ha3a6a3e_1
libiconv=1.16=h1a28f6b_2
liblapack=3.9.0=17_osxarm64_openblas
liblapacke=3.9.0=17_osxarm64_openblas
libnghttp2=1.52.0=h62f6fdd_1
libopenblas=0.3.23=openmp_hc731615_0
libpng=1.6.39=h80987f9_0
libprotobuf=3.20.3=h514c7bf_0
libssh2=1.10.0=h02f6b3c_2
libthrift=0.15.0=h73c2103_2
libtiff=4.5.0=h313beb8_2
libuv=1.44.2=h80987f9_0
libwebp=1.2.4=ha3663a8_1
libwebp-base=1.2.4=h80987f9_1
llvm-openmp=16.0.6=h1c12783_0
lz4-c=1.9.4=h313beb8_0
markupsafe=2.1.1=py311h80987f9_0
mpc=1.1.0=h8c48613_1
mpfr=4.0.2=h695f6f0_1
mpmath=1.3.0=py311hca03da5_0
multidict=6.0.2=py311h80987f9_0
multiprocess=0.70.14=py311hca03da5_0
murmurhash=1.0.7=py311h313beb8_0
ncurses=6.4=h313beb8_0
networkx=3.1=py311hca03da5_0
ninja=1.10.2=hca03da5_5
ninja-base=1.10.2=h525c30c_5
numexpr=2.8.4=py311h6dc990b_1
numpy=1.25.2=py311he598dae_0
numpy-base=1.25.2=py311hfbfe69c_0
openblas=0.3.23=openmp_hf78f355_0
openssl=3.1.2=h53f4e23_0
orc=1.7.4=hdca1487_1
packaging=23.1=pyhd8ed1ab_0
pandas=1.5.3=py311h6956b77_0
pathy=0.10.2=pyhd8ed1ab_0
pillow=9.4.0=py311h313beb8_0
pip=23.2.1=py311hca03da5_0
preshed=3.0.6=py311h313beb8_0
pyarrow=11.0.0=py311h7575258_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.8=py311h80987f9_0
pygments=2.16.1=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.11.4=hb885b13_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-xxhash=2.0.2=py311h80987f9_1
pytorch=2.0.1=py3.11_0
pytz=2022.7=py311hca03da5_0
pyyaml=6.0=py311h80987f9_1
re2=2022.04.01=hc377ac9_0
readline=8.2=h1a28f6b_0
regex=2022.7.9=py311h80987f9_0
requests=2.31.0=pyhd8ed1ab_0
rich=12.6.0=pyhd8ed1ab_0
sacremoses=master=py_0
setuptools=68.0.0=py311hca03da5_0
shellingham=1.5.1=pyhd8ed1ab_0
six=1.16.0=pyhd3eb1b0_1
sleef=3.5.1=h80987f9_2
smart_open=5.2.1=pyhd8ed1ab_0
snappy=1.1.9=hc377ac9_0
spacy=3.5.3=py311h30ceab6_0
spacy-legacy=3.0.12=pyhd8ed1ab_0
spacy-loggers=1.0.4=pyhd8ed1ab_0
sqlite=3.41.2=h80987f9_0
srsly=2.4.6=py311h313beb8_0
sympy=1.11.1=py311hca03da5_0
thinc=8.1.10=py311h30ceab6_0
tk=8.6.12=hb8d0fd4_0
tokenizers=0.13.2=py311h3dd52b7_1
torchaudio=2.0.2=py311_cpu
torchvision=0.15.2=cpu_py311he74fb5d_0
tqdm=4.66.1=pyhd8ed1ab_0
transformers=4.29.2=py311hca03da5_0
typer=0.7.0=pyhd8ed1ab_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023c=h04d1e81_0
urllib3=2.0.4=pyhd8ed1ab_0
utf8proc=2.6.1=h1a28f6b_0
wasabi=0.10.1=pyhd8ed1ab_1
wheel=0.38.4=py311hca03da5_0
xxhash=0.8.0=h1a28f6b_3
xz=5.4.2=h80987f9_0
yaml=0.2.5=h1a28f6b_0
yarl=1.8.1=py311h80987f9_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.5=hd90d995_0
